Extract upload workflow hooks and subcomponents from UploadPanel
UploadPanel (575 lines) becomes a ~220-line composition. Form state
with codec bitrate-default migration moves to useUploadForm; the
debounced, abortable price quote to useUploadQuote; multipart
submission with progress to useUploadSubmit. The dropzone (owning its
drag state), encode settings + rendition grid, and quote summary
become upload/{FileDropZone,EncodeSettingsFields,UploadQuoteSummary}.
Upload test suite passes unchanged.
